cleanup and unifying the coding style
All indentations are now 1 tab instead of a random number of spaces.
We can technically still change this but 2 out of 3 prefer tabs over
spaces. I did not touch the html file because that is a mess with even
more random indentations and I'm not sure how intentional that is.