Complete rework of ourglass
Usage of ourlgass is just as easy as before and also recieved an update. Now
there is a reset functionallity to reset the ourglass back to the initial
state.

The logic is mainly rework, as now the creation of the single elements is done
dynamically. Only selected phases are created and animated. This is achieved by
validating the input from the ourglass form and creating elements inside the
java script. Heights of the single triangles are calculated, that the visible
area is proportional to the phase time.

And the best thing is: New Animation curves for smoooooth transition between
phases.

- send from a train somewhere in lower Saxony